+++ src/ThreeSceneSetup/HexGrid/HexGrid.ts
@@ -4,16 +4,18 @@ import HexTile from './HexTile';
 class HexGrid {
   radius: number;
   tileSize: number;
+  private readonly color: string;
 
-  constructor(radius: number, tileSize: number) {
+  constructor(radius: number, tileSize: number, color: string) {
     this.radius = radius;
     this.tileSize = tileSize;
+    this.color = color;
   }
 
   generateGrid(): HexTile[] {
     const hexTiles: HexTile[] = [];
-    const width = Math.sqrt(3) * this.tileSize;  // Горизонтальное расстояние между центрами шестиугольников
-    const height = 2 * this.tileSize;            // Вертикальное расстояние между центрами шестиугольников
+    const width = Math.sqrt(3) * this.tileSize;
+    const height = 2 * this.tileSize;
 
     for (let q = -this.radius; q <= this.radius; q++) {
       const r1 = Math.max(-this.radius, -q - this.radius);
@@ -21,7 +23,7 @@ class HexGrid {
       for (let r = r1; r <= r2; r++) {
         const x = width * (q + r / 2);
         const z = height * 0.75 * r;
-        const tile = new HexTile(new THREE.Vector3(x, -0.1, z), this.tileSize);
+        const tile = new HexTile(new THREE.Vector3(x, -0.1, z), this.tileSize, this.color);
         hexTiles.push(tile);
       }
     }